Update Intake_exhaust.mlx
Added the ability to plot PV diagram at the end. The PV diagram is surprising, for I always thought that intake or exhaust happens at constant pressure.
There is an issue, however: The function V(t) is not vectorized and so I worked around by plotting each point separately. The execution time is considerable. Added to issues, and looking for someone to fix the problem.